Testing complete
- Fixes in coloning
- Support fo cntrl+c & cntrl+v combinations
