+++ src/main/java/tf/gpx/edit/helper/GPXListHelper.java
@@ -25,6 +25,7 @@
  */
 package tf.gpx.edit.helper;
 
+import tf.gpx.edit.items.GPXLineItem;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
