+++ src/main/java/tf/gpx/edit/items/GPXTrackSegment.java
@@ -23,15 +23,13 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package tf.gpx.edit.helper;
+package tf.gpx.edit.items;
 
 import com.hs.gpxparser.modal.Extension;
-import com.hs.gpxparser.modal.GPX;
-import com.hs.gpxparser.modal.Route;
+import com.hs.gpxparser.modal.Track;
 import com.hs.gpxparser.modal.TrackSegment;
 import com.hs.gpxparser.modal.Waypoint;
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.Date;
 import java.util.LinkedList;
 import java.util.List;
@@ -39,15 +37,16 @@ import java.util.Set;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.geometry.BoundingBox;
-import static tf.gpx.edit.helper.GPXLineItem.filterGPXWaypointsInBoundingBox;
+import tf.gpx.edit.helper.EarthGeometry;
+import tf.gpx.edit.helper.GPXCloner;
 
 /**
  *
  * @author Thomas
  */
-public class GPXRoute extends GPXMeasurable {
-    private GPXFile myGPXFile;
-    private Route myRoute;
+public class GPXTrackSegment extends GPXMeasurable {
+    private GPXTrack myGPXTrack;
+    private TrackSegment myTrackSegment;
     private final ObservableList<GPXWaypoint> myGPXWaypoints = FXCollections.observableList(new LinkedList<>());
     
     private Double myLength = null;
@@ -55,42 +54,48 @@ public class GPXRoute extends GPXMeasurable {
     private Double myCumulativeDescent = null;
     private Double myMinHeight = null;
     private Double myMaxHeight = null;
+    private Date myStartingTime = null;
+    private Date myEndTime = null;
     
-    private GPXRoute() {
-        super(GPXLineItemType.GPXRoute);
+    private GPXTrackSegment() {
+        super(GPXLineItemType.GPXTrackSegment);
     }
     
-    // constructor for "manually created routes"
-    public GPXRoute(final GPXFile gpxFile) {
-        super(GPXLineItemType.GPXRoute);
+    // constructor for "manually created tracksegments"
+    public GPXTrackSegment(final GPXTrack gpxTrack) {
+        super(GPXLineItemType.GPXTrackSegment);
 
-        myGPXFile = gpxFile;
+        myGPXTrack = gpxTrack;
 
-        // create empty route
-        myRoute = new Route();
+        // create tracksegment
+        myTrackSegment = new TrackSegment();
         
-        // if possible add route to parent class
-        Extension content = gpxFile.getContent();
-        if (content instanceof GPX) {
-            ((GPX) content).addRoute(myRoute);
+        // if possible add waypoint to parent class
+        Extension content = gpxTrack.getContent();
+        if (content instanceof Track) {
+            ((Track) content).addTrackSegment(myTrackSegment);
         }
         
         myGPXWaypoints.addListener(getListChangeListener());
     }
     
-    // constructor for routes from gpx parser
-    public GPXRoute(final GPXFile gpxFile, final Route route) {
-        super(GPXLineItemType.GPXRoute);
+    // constructor for tracksegments from gpx parser
+    public GPXTrackSegment(
+            final GPXTrack gpxTrack, 
+            final TrackSegment trackSegment, 
+            final int number) {
+        super(GPXLineItemType.GPXTrackSegment);
         
-        myGPXFile = gpxFile;
-        myRoute = route;
+        myGPXTrack = gpxTrack;
+        myTrackSegment = trackSegment;
+        setNumber(number);
         
         // TFE, 20180203: tracksegment without wayoints is valid!
-        if (myRoute.getRoutePoints() != null) {
-            for (Waypoint waypoint : myRoute.getRoutePoints()) {
+        if (myTrackSegment.getWaypoints() != null) {
+            for (Waypoint waypoint : myTrackSegment.getWaypoints()) {
                 myGPXWaypoints.add(new GPXWaypoint(this, waypoint, myGPXWaypoints.size()+1));
             }
-            assert (myGPXWaypoints.size() == myRoute.getRoutePoints().size());
+            assert (myGPXWaypoints.size() == myTrackSegment.getWaypoints().size());
 
             updatePrevNextGPXWaypoints();
         }
@@ -99,19 +104,31 @@ public class GPXRoute extends GPXMeasurable {
     }
     
     @Override
-    public GPXRoute cloneMeWithChildren() {
-        final GPXRoute myClone = new GPXRoute();
+    public GPXTrackSegment cloneMeWithChildren() {
+        final GPXTrackSegment myClone = new GPXTrackSegment();
         
         // parent needs to be set initially - list functions use this for checking
-        myClone.myGPXFile = myGPXFile;
+        myClone.myGPXTrack = myGPXTrack;
         
-        // set route via cloner
-        myClone.myRoute = GPXCloner.getInstance().deepClone(myRoute);
+        // set tracksegment via cloner
+        myClone.myTrackSegment = GPXCloner.getInstance().deepClone(myTrackSegment);
+        
+        myClone.myLength = myLength;
+        myClone.myCumulativeAscent = myCumulativeAscent;
+        myClone.myCumulativeDescent = myCumulativeDescent;
+        myClone.myMinHeight = myMinHeight;
+        myClone.myMaxHeight = myMaxHeight;
+        myClone.myStartingTime = myStartingTime;
+        myClone.myEndTime = myEndTime;
         
         // clone all my children
         for (GPXWaypoint gpxWaypoint : myGPXWaypoints) {
             myClone.myGPXWaypoints.add(gpxWaypoint.cloneMeWithChildren());
         }
+        numberChildren(myClone.myGPXWaypoints);
+
+        // init prev/next waypoints
+        myClone.updatePrevNextGPXWaypoints();
 
         myClone.myGPXWaypoints.addListener(getListChangeListener());
 
@@ -119,20 +136,20 @@ public class GPXRoute extends GPXMeasurable {
         return myClone;
     }
 
-    protected Route getRoute() {
-        return myRoute;
+    protected TrackSegment getTrackSegment() {
+        return myTrackSegment;
     }
     
     @Override
     public GPXLineItem getParent() {
-        return myGPXFile;
+        return myGPXTrack;
     }
 
     @Override
-    public void setParent(final GPXLineItem parent) {
-        assert GPXLineItem.GPXLineItemType.GPXFile.equals(parent.getType());
+    public void setParent(GPXLineItem parent) {
+        assert GPXLineItem.GPXLineItemType.GPXTrack.equals(parent.getType());
         
-        myGPXFile = (GPXFile) parent;
+        myGPXTrack = (GPXTrack) parent;
         setHasUnsavedChanges();
     }
 
@@ -155,6 +172,8 @@ public class GPXRoute extends GPXMeasurable {
         myLength = null;
         myCumulativeAscent = null;
         myCumulativeDescent = null;
+        myStartingTime = null;
+        myEndTime = null;
         
         setHasUnsavedChanges();
     }
@@ -175,28 +194,12 @@ public class GPXRoute extends GPXMeasurable {
     }
     
     @Override
-    public Integer getNumber() {
-        return myRoute.getNumber();
-    }
-
-    @Override
-    public void setNumber(Integer number) {
-        myRoute.setNumber(number);
-        setHasUnsavedChanges();
-    }
-
-    public Comparator<GPXRoute> getComparator() {
-        return (GPXRoute a, GPXRoute b) -> a.getNumber() - b.getNumber();
-    }
-    
-    @Override
     public String getName() {
-        return myRoute.getName();
+        return myGPXTrack.getName() + " - Segment " + getNumber();
     }
 
     @Override
     public void setName(final String name) {
-        myRoute.setName(name);
     }
     
     @Override
@@ -206,30 +209,31 @@ public class GPXRoute extends GPXMeasurable {
     
     @Override
     public Date getDate() {
-        return null;
+        return getStartTime();
     }
 
     @Override
     public GPXFile getGPXFile() {
-        return myGPXFile;
+        return getParent().getGPXFile();
     }
 
     @Override
     public ObservableList<GPXTrack> getGPXTracks() {
         ObservableList<GPXTrack> result = FXCollections.observableArrayList();
+        result.add(myGPXTrack);
         return result;
     }
 
     @Override
     public ObservableList<GPXTrackSegment> getGPXTrackSegments() {
         ObservableList<GPXTrackSegment> result = FXCollections.observableArrayList();
+        result.add(this);
         return result;
     }
 
     @Override
     public ObservableList<GPXRoute> getGPXRoutes() {
         ObservableList<GPXRoute> result = FXCollections.observableArrayList();
-        result.add(this);
         return result;
     }
 
@@ -240,13 +244,13 @@ public class GPXRoute extends GPXMeasurable {
     
     @Override
     public Extension getContent() {
-        return myRoute;
+        return myTrackSegment;
     }
 
     @Override
     public ObservableList<GPXWaypoint> getCombinedGPXWaypoints(final GPXLineItemType itemType) {
         ObservableList<GPXWaypoint> result = FXCollections.observableArrayList();
-        if (itemType == null || itemType.equals(GPXLineItemType.GPXRoute)) {
+        if (itemType == null || itemType.equals(GPXLineItemType.GPXTrack) || itemType.equals(GPXLineItemType.GPXTrackSegment)) {
             result = myGPXWaypoints;
         }
         return result;
@@ -260,7 +264,7 @@ public class GPXRoute extends GPXMeasurable {
     /**
      * Calculates the getLength of the track segment
      * 
-     * @return the route's getLength in meters
+     * @return the segment's getLength in meters
      */
     @Override
     public double getLength() {
@@ -270,16 +274,15 @@ public class GPXRoute extends GPXMeasurable {
         
         double length = 0.0;
 
-        GPXWaypoint currentWaypoint;
-        GPXWaypoint previousWaypoint;
-
+        GPXWaypoint previousWaypoint = null;
         /* Only attempt to calculate the distanceGPXWaypoints if we are not
          * on the first way point of the segment. */
-        for (int z = 1; z < myGPXWaypoints.size(); z++) {
-            currentWaypoint = myGPXWaypoints.get(z);
-            previousWaypoint = myGPXWaypoints.get(z - 1);
-
-            length += EarthGeometry.distanceGPXWaypoints(currentWaypoint, previousWaypoint);
+        for (GPXWaypoint gpxWaypoint : myGPXWaypoints) {
+            if (previousWaypoint != null) {
+                length += EarthGeometry.distanceGPXWaypoints(gpxWaypoint, previousWaypoint);
+            }
+            
+            previousWaypoint = gpxWaypoint;
         }
 
         myLength = length;
@@ -290,12 +293,12 @@ public class GPXRoute extends GPXMeasurable {
      * Calculates the total ascent in the segment.
      * 
      * <p>The total ascent of the segment is calculated by comparing each
-     * of the route's way point with their predecessors. If the
+     * of the segment's way point with their predecessors. If the
      * elevation of a way point is higher than the elevation of the
      * predecessor, the total ascent is increased accordingly.</p>
      * 
      * @see TrackSegment#cumulativeDescent()
-     * @return the route's total ascent in meters
+     * @return the segment's total ascent in meters
      */
     @Override
     public double getCumulativeAscent() {
@@ -304,14 +307,15 @@ public class GPXRoute extends GPXMeasurable {
         }
         double ascent = 0.0;
 
-        if (myGPXWaypoints.size() <= 1) {
-            return 0.0;
-        }
-
-        for (int i = 0; i < myGPXWaypoints.size(); i++) {
-            if (i > 0 && myGPXWaypoints.get(i - 1).getWaypoint().getElevation() < myGPXWaypoints.get(i).getWaypoint().getElevation()) {
-                ascent += myGPXWaypoints.get(i).getWaypoint().getElevation() - myGPXWaypoints.get(i - 1).getWaypoint().getElevation();
+        GPXWaypoint previousWaypoint = null;
+        /* Only attempt to calculate the distanceGPXWaypoints if we are not
+         * on the first way point of the segment. */
+        for (GPXWaypoint gpxWaypoint : myGPXWaypoints) {
+            if ((previousWaypoint != null) && (previousWaypoint.getWaypoint().getElevation() < gpxWaypoint.getWaypoint().getElevation())) {
+                ascent += gpxWaypoint.getWaypoint().getElevation() - previousWaypoint.getWaypoint().getElevation();
             }
+            
+            previousWaypoint = gpxWaypoint;
         }
 
         myCumulativeAscent = ascent;
@@ -322,11 +326,11 @@ public class GPXRoute extends GPXMeasurable {
      * Calculates the total descent in the segment.
      * 
      * <p>The total descent of the segment is calculated by comparing each
-     * of the route's way point with their predecessors. If the
+     * of the segment's way point with their predecessors. If the
      * elevation of a way point is lower than the elevation of the
      * predecessor, the total descent is increased accordingly.</p>
      * 
-     * @return the route's total descent in meters
+     * @return the segment's total descent in meters
      * 
      * @see TrackSegment#cumulativeAscent()
      */
@@ -338,14 +342,15 @@ public class GPXRoute extends GPXMeasurable {
 
         double descent = 0.0;
 
-        if (myGPXWaypoints.size() <= 1) {
-            return 0.0;
-        }
-
-        for (int i = 0; i < myGPXWaypoints.size(); i++) {
-            if (i > 1 && myGPXWaypoints.get(i).getWaypoint().getElevation() < myGPXWaypoints.get(i - 1).getWaypoint().getElevation()) {
-                descent += myGPXWaypoints.get(i - 1).getWaypoint().getElevation() - myGPXWaypoints.get(i).getWaypoint().getElevation();
+        GPXWaypoint previousWaypoint = null;
+        /* Only attempt to calculate the distanceGPXWaypoints if we are not
+         * on the first way point of the segment. */
+        for (GPXWaypoint gpxWaypoint : myGPXWaypoints) {
+            if ((previousWaypoint != null) && (gpxWaypoint.getWaypoint().getElevation() < previousWaypoint.getWaypoint().getElevation())) {
+                descent += previousWaypoint.getWaypoint().getElevation() - gpxWaypoint.getWaypoint().getElevation();
             }
+            
+            previousWaypoint = gpxWaypoint;
         }
 
         myCumulativeDescent = descent;
@@ -353,27 +358,69 @@ public class GPXRoute extends GPXMeasurable {
     }
 
     /**
-     * Returns null since route waypoints have no time
+     * Returns the point in time when the segment was entered
      * 
-     * @return null 
+     * <p>Usually this is the time stamp of the way point that was added
+     * first to the segment.</p>
+     * 
+     * @see TrackSegment#endTime
+     * @return the point in time when the segment was entered 
      */
     @Override
     protected Date getStartTime() {
-        return null;
+        if (myStartingTime != null) {
+            return myStartingTime;
+        }
+
+        Date result = null;
+
+        for (GPXWaypoint gpxWaypoint : myGPXWaypoints) {
+            Date time = gpxWaypoint.getWaypoint().getTime();
+
+            if (time != null) {
+                if (result == null || time.before(result)) {
+                    result = time;
+                }
+            }
+        }
+
+        myStartingTime = result;
+        return result;
     }
 
     /**
-     * Returns null since route waypoints have no time
+     * Returns the point in time when the segment was left
      * 
-     * @return null 
+     * <p>Usually this is the time stamp of the way point that was added
+     * last to the segment.</p>
+     *
+     * @see TrackSegment#endTime
+     * @return the point in time when the segment was left
      */
     @Override
     protected Date getEndTime() {
-        return null;
+        if (myEndTime != null) {
+            return myEndTime;
+        }
+
+        Date result = null;
+
+        for (GPXWaypoint gpxWaypoint : myGPXWaypoints) {
+            Date time = gpxWaypoint.getWaypoint().getTime();
+
+            if (time != null) {
+                if (result == null || time.after(result)) {
+                    result = time;
+                }
+            }
+        }
+
+        myEndTime = result;
+        return result;
     }
     
     /**
-     * @return the minimum height of the route
+     * @return the minimum height of the track
      */
     @Override
     public double getMinHeight() {
@@ -383,9 +430,9 @@ public class GPXRoute extends GPXMeasurable {
 
         double result = Double.MAX_VALUE;
 
-        for (int i = 0; i < myGPXWaypoints.size(); i++) {
-            if (myGPXWaypoints.get(i).getWaypoint().getElevation() < result) {
-                result = myGPXWaypoints.get(i).getWaypoint().getElevation();
+        for (GPXWaypoint gpxWaypoint : myGPXWaypoints) {
+            if (gpxWaypoint.getWaypoint().getElevation() < result) {
+                result = gpxWaypoint.getWaypoint().getElevation();
             }
         }
 
@@ -394,7 +441,7 @@ public class GPXRoute extends GPXMeasurable {
     }
     
     /**
-     * @return the maximum height of the route
+     * @return the maximum height of the track
      */
     @Override
     public double getMaxHeight() {
@@ -404,9 +451,9 @@ public class GPXRoute extends GPXMeasurable {
 
         double result = Double.MIN_VALUE;
 
-        for (int i = 0; i < myGPXWaypoints.size(); i++) {
-            if (myGPXWaypoints.get(i).getWaypoint().getElevation() > result) {
-                result = myGPXWaypoints.get(i).getWaypoint().getElevation();
+        for (GPXWaypoint gpxWaypoint : myGPXWaypoints) {
+            if (gpxWaypoint.getWaypoint().getElevation() > result) {
+                result = gpxWaypoint.getWaypoint().getElevation();
             }
         }
 
@@ -416,7 +463,7 @@ public class GPXRoute extends GPXMeasurable {
 
     @Override
     protected void visitMe(final IGPXLineItemVisitor visitor) {
-        visitor.visitGPXRoute(this);
+        visitor.visitGPXTrackSegment(this);
     }
 
     @Override
@@ -427,7 +474,7 @@ public class GPXRoute extends GPXMeasurable {
             });
             
             final Set<Waypoint> waypoints = numberExtensions(myGPXWaypoints);
-            myRoute.setRoutePoints(new ArrayList<>(waypoints));
+            myTrackSegment.setWaypoints(new ArrayList<>(waypoints));
 
             updatePrevNextGPXWaypoints();
 
@@ -435,6 +482,8 @@ public class GPXRoute extends GPXMeasurable {
             myLength = null;
             myCumulativeAscent = null;
             myCumulativeDescent = null;
+            myStartingTime = null;
+            myEndTime = null;
         }
     }
 }
