+++ src/main/java/tf/gpx/edit/kml/KMLWriter.java
@@ -42,7 +42,7 @@ import javax.xml.transform.stream.StreamResult;
 import org.w3c.dom.DOMException;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
-import tf.gpx.edit.helper.GPXWaypoint;
+import tf.gpx.edit.items.GPXWaypoint;
 import tf.gpx.edit.viewer.MarkerManager;
 
 /**
