+++ src/main/java/tf/gpx/edit/main/GPXEditor.java
@@ -112,14 +112,14 @@ import tf.gpx.edit.values.EditGPXMetadata;
 import tf.gpx.edit.helper.GPXEditorParameters;
 import tf.gpx.edit.helper.GPXEditorPreferences;
 import tf.gpx.edit.helper.GPXEditorWorker;
-import tf.gpx.edit.helper.GPXFile;
-import tf.gpx.edit.helper.GPXLineItem;
-import tf.gpx.edit.helper.GPXRoute;
-import tf.gpx.edit.helper.GPXTrack;
-import tf.gpx.edit.helper.GPXTrackSegment;
+import tf.gpx.edit.items.GPXFile;
+import tf.gpx.edit.items.GPXLineItem;
+import tf.gpx.edit.items.GPXRoute;
+import tf.gpx.edit.items.GPXTrack;
+import tf.gpx.edit.items.GPXTrackSegment;
 import tf.gpx.edit.viewer.GPXTrackviewer;
 import tf.gpx.edit.helper.GPXTreeTableView;
-import tf.gpx.edit.helper.GPXWaypoint;
+import tf.gpx.edit.items.GPXWaypoint;
 import tf.gpx.edit.parser.DefaultExtensionHolder;
 import tf.gpx.edit.srtm.AssignSRTMHeight;
 import tf.gpx.edit.srtm.SRTMDataStore;
