+++ src/main/java/tf/gpx/edit/values/EditGPXWaypoint.java
@@ -54,8 +54,8 @@ import jfxtras.labs.scene.control.BigDecimalField;
 import jfxtras.scene.control.CalendarTextField;
 import tf.gpx.edit.general.RestrictiveTextField;
 import tf.gpx.edit.general.TooltipHelper;
-import tf.gpx.edit.helper.GPXLineItem;
-import tf.gpx.edit.helper.GPXWaypoint;
+import tf.gpx.edit.items.GPXLineItem;
+import tf.gpx.edit.items.GPXWaypoint;
 import tf.gpx.edit.helper.LatLongHelper;
 import tf.gpx.edit.main.GPXEditor;
 import tf.gpx.edit.main.GPXEditorManager;
