+++ src/main/java/tf/gpx/edit/viewer/MarkerManager.java
@@ -31,7 +31,7 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import tf.gpx.edit.helper.GPXWaypoint;
+import tf.gpx.edit.items.GPXWaypoint;
 
 /**
  * One manager to role them all...
